mlpr/requirements.txt
google-cloud-vision
google-cloud-translate
streamlit
Pillow
